requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap-datepicker-plus==3.0.5
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-heroku==0.0.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
pbr==5.5.1
Pillow==8.0.1
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stevedore==3.3.0
urllib3==1.26.2
virtualenv==20.3.1
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whitenoise==5.2.0
